fix(MacOS): Allow Mac users more system keybinds
Online doesn't know what keybinds it has. Instead, it leaves knowing
that up to core. This means we preventDefault() on keybinds relatively
indiscriminately, having explicit carve-outs for keybinds we want to
execute - such as Ctrl+Shift+I to open the browser devtools, and sending
everything to core.

Sometimes, we may have keybinds which don't have a core equivalent - for
example Cmd+Ctrl+Space. In this case the keybind is ignored in core and
Everything Is Fine... mostly. The problem is when these bindings collide
with system bindings. Cmd+Ctrl+Space happens to be the binding on MacOS
to open the emoji picker... which previously never opened in Collabora
Online. Another example is the MacOS minimize keybind, Cmd+M, which
appeared to do nothing but was preventing a MacOS system shortcut in
some browsers.

Similarly, sometimes there are keybinds which, while interpreted as some
action in core, have a well-understood meaning for the system. An
example is Cmd+R or Cmd+M (refresh tab, used in LibreOffice as "justify
text right"). This is such an obviously-bad thing to bind over that
Safari won't let you!.. but Chrome and Firefox both will, causing some
nasty browser inconsistencies and breaking users' expectations of how
the web works.

Don't do that.

Signed-off-by: Skyler Grey <skyler.grey@collabora.com>
Change-Id: I3b5c3390e5eb88e8e38e552fa424cedd946f65e9